Fixed few thing in api/api_filters.c. Fixed calculation of Gyro rate parameters.

+++ src/api/api_filters.c
@@ -29,28 +29,31 @@ z = 0.884414 + j 0.209794
 #define NCoef 4
 float float_api_filters_iir_acc_x(float float_new_sample)
 {
-    double ACoef[NCoef+1] = {
-        0.00018539265053394222,
-        0.00074157060213576886,
-        0.00111235590320365330,
-        0.00074157060213576886,
-        0.00018539265053394222
+    float ACoef[NCoef+1] = {
+        0.00018539265053394222f,
+        0.00074157060213576886f,
+        0.00111235590320365330f,
+        0.00074157060213576886f,
+        0.00018539265053394222f
     };
 
-    double BCoef[NCoef+1] = {
-        1.00000000000000000000,
-        -3.34406783771187310000,
-        4.23886395088406330000,
-        -2.40934285658631710000,
-        0.51747819978804011000
+    float BCoef[NCoef+1] = {
+        1.00000000000000000000f,
+        -3.34406783771187310000f,
+        4.23886395088406330000f,
+        -2.40934285658631710000f,
+        0.51747819978804011000f
     };
-
-    static float y[NCoef+1]; /* Output samples. */
-    static float x[NCoef+1]; /* Input samples.  */
+    
+    /* Output samples. */
+    static float y[NCoef+1]; 
+    
+    /* Input samples.  */
+    static float x[NCoef+1]; 
     int n;
 
     /* shift the old samples */
-    for(n=NCoef; n>0; n--)
+    for(n = NCoef; n > 0; n--)
     {
        x[n] = x[n-1];
        y[n] = y[n-1];
@@ -58,38 +61,41 @@ float float_api_filters_iir_acc_x(float float_new_sample)
 
     /* Calculate the new output */
     x[0] = float_new_sample;
-    y[0] = (float)ACoef[0] * x[0];
-    for(n=1; n<=NCoef; n++)
+    y[0] = ACoef[0] * x[0];
+    for(n = 1; n <= NCoef; n++)
     {
-        y[0] += (float)ACoef[n] * x[n] - (float)BCoef[n] * y[n];
+        y[0] += ACoef[n] * x[n] - BCoef[n] * y[n];
     }
     return y[0];
 }
 
 float float_api_filters_iir_acc_y(float float_new_sample)
 {
-    double ACoef[NCoef+1] = {
-        0.00018539265053394222,
-        0.00074157060213576886,
-        0.00111235590320365330,
-        0.00074157060213576886,
-        0.00018539265053394222
+    float ACoef[NCoef+1] = {
+        0.00018539265053394222f,
+        0.00074157060213576886f,
+        0.00111235590320365330f,
+        0.00074157060213576886f,
+        0.00018539265053394222f
     };
 
-    double BCoef[NCoef+1] = {
-        1.00000000000000000000,
-        -3.34406783771187310000,
-        4.23886395088406330000,
-        -2.40934285658631710000,
-        0.51747819978804011000
+    float BCoef[NCoef+1] = {
+        1.00000000000000000000f,
+        -3.34406783771187310000f,
+        4.23886395088406330000f,
+        -2.40934285658631710000f,
+        0.51747819978804011000f
     };
-
-    static float y[NCoef+1]; /* Output samples. */
-    static float x[NCoef+1]; /* Input samples.  */
+    
+    /* Output samples. */
+    static float y[NCoef+1]; 
+    
+    /* Input samples.  */
+    static float x[NCoef+1]; 
     int n;
 
     /* shift the old samples */
-    for(n=NCoef; n>0; n--)
+    for(n = NCoef; n > 0; n--)
     {
        x[n] = x[n-1];
        y[n] = y[n-1];
@@ -97,38 +103,41 @@ float float_api_filters_iir_acc_y(float float_new_sample)
 
     /* Calculate the new output */
     x[0] = float_new_sample;
-    y[0] = (float)ACoef[0] * x[0];
-    for(n=1; n<=NCoef; n++)
+    y[0] = ACoef[0] * x[0];
+    for(n = 1; n <= NCoef; n++)
     {
-        y[0] += (float)ACoef[n] * x[n] - (float)BCoef[n] * y[n];
+        y[0] += ACoef[n] * x[n] - BCoef[n] * y[n];
     }
     return y[0];
 }
 
 float float_api_filters_iir_acc_z(float float_new_sample)
 {
-    double ACoef[NCoef+1] = {
-        0.00018539265053394222,
-        0.00074157060213576886,
-        0.00111235590320365330,
-        0.00074157060213576886,
-        0.00018539265053394222
+    float ACoef[NCoef+1] = {
+        0.00018539265053394222f,
+        0.00074157060213576886f,
+        0.00111235590320365330f,
+        0.00074157060213576886f,
+        0.00018539265053394222f
     };
 
-    double BCoef[NCoef+1] = {
-        1.00000000000000000000,
-        -3.34406783771187310000,
-        4.23886395088406330000,
-        -2.40934285658631710000,
-        0.51747819978804011000
+    float BCoef[NCoef+1] = {
+        1.00000000000000000000f,
+        -3.34406783771187310000f,
+        4.23886395088406330000f,
+        -2.40934285658631710000f,
+        0.51747819978804011000f
     };
-
-    static float y[NCoef+1]; /* Output samples. */
-    static float x[NCoef+1]; /* Input samples.  */
+    
+    /* Output samples. */
+    static float y[NCoef+1]; 
+    
+    /* Input samples.  */
+    static float x[NCoef+1]; 
     int n;
 
     /* shift the old samples */
-    for(n=NCoef; n>0; n--)
+    for(n = NCoef; n > 0; n--)
     {
        x[n] = x[n-1];
        y[n] = y[n-1];
@@ -136,10 +145,10 @@ float float_api_filters_iir_acc_z(float float_new_sample)
 
     /* Calculate the new output */
     x[0] = float_new_sample;
-    y[0] = (float)ACoef[0] * x[0];
-    for(n=1; n<=NCoef; n++)
+    y[0] = ACoef[0] * x[0];
+    for(n = 1; n <= NCoef; n++)
     {
-        y[0] += (float)ACoef[n] * x[n] - (float)BCoef[n] * y[n];
+        y[0] += ACoef[n] * x[n] - BCoef[n] * y[n];
     }
     return y[0];
 }
@@ -172,28 +181,30 @@ z = 0.369637 + j 0.603107
 #define NCoef 4
 float float_api_filters_iir_gyro_x(float float_new_sample)
 {
-    double ACoef[NCoef+1] = {
-        0.02675265688403948700,
-        0.10701062753615795000,
-        0.16051594130423694000,
-        0.10701062753615795000,
-        0.02675265688403948700
+    float ACoef[NCoef+1] = {
+        0.02675265688403948700f,
+        0.10701062753615795000f,
+        0.16051594130423694000f,
+        0.10701062753615795000f,
+        0.02675265688403948700f
     };
 
-    double BCoef[NCoef+1] = {
-        1.00000000000000000000,
-        -1.28555068801461640000,
-        1.01289917136114170000,
-        -0.35368582226312562000,
-        0.05438080336347629600
+    float BCoef[NCoef+1] = {
+        1.00000000000000000000f,
+        -1.28555068801461640000f,
+        1.01289917136114170000f,
+        -0.35368582226312562000f,
+        0.05438080336347629600f
     };
-
-    static float y[NCoef+1]; /* output samples*/
-    static float x[NCoef+1]; /* input samples */
+    
+    /* output samples*/
+    static float y[NCoef+1]; 
+    /* input samples */
+    static float x[NCoef+1]; 
     int n;
 
     /* shift the old samples */
-    for(n=NCoef; n>0; n--)
+    for(n = NCoef; n>0; n--)
     {
        x[n] = x[n-1];
        y[n] = y[n-1];
@@ -201,38 +212,40 @@ float float_api_filters_iir_gyro_x(float float_new_sample)
 
     /* Calculate the new output */
     x[0] = float_new_sample;
-    y[0] = (float)ACoef[0] * x[0];
-    for(n=1; n<=NCoef; n++)
+    y[0] = ACoef[0] * x[0];
+    for(n = 1; n <= NCoef; n++)
     {
-        y[0] += (float)ACoef[n] * x[n] - (float)BCoef[n] * y[n];
+        y[0] += ACoef[n] * x[n] - BCoef[n] * y[n];
     }
     return y[0];
 }
 
 float float_api_filters_iir_gyro_y(float float_new_sample)
 {
-    double ACoef[NCoef+1] = {
-        0.02675265688403948700,
-        0.10701062753615795000,
-        0.16051594130423694000,
-        0.10701062753615795000,
-        0.02675265688403948700
+    float ACoef[NCoef+1] = {
+        0.02675265688403948700f,
+        0.10701062753615795000f,
+        0.16051594130423694000f,
+        0.10701062753615795000f,
+        0.02675265688403948700f
     };
 
-    double BCoef[NCoef+1] = {
-        1.00000000000000000000,
-        -1.28555068801461640000,
-        1.01289917136114170000,
-        -0.35368582226312562000,
-        0.05438080336347629600
+    float BCoef[NCoef+1] = {
+        1.00000000000000000000f,
+        -1.28555068801461640000f,
+        1.01289917136114170000f,
+        -0.35368582226312562000f,
+        0.05438080336347629600f
     };
-
-    static float y[NCoef+1]; /* output samples*/
-    static float x[NCoef+1]; /* input samples */
+    
+    /* output samples*/
+    static float y[NCoef+1]; 
+    /* input samples */
+    static float x[NCoef+1]; 
     int n;
 
     /* shift the old samples */
-    for(n=NCoef; n>0; n--)
+    for(n = NCoef; n>0; n--)
     {
        x[n] = x[n-1];
        y[n] = y[n-1];
@@ -240,38 +253,40 @@ float float_api_filters_iir_gyro_y(float float_new_sample)
 
     /* Calculate the new output */
     x[0] = float_new_sample;
-    y[0] = (float)ACoef[0] * x[0];
-    for(n=1; n<=NCoef; n++)
+    y[0] = ACoef[0] * x[0];
+    for(n = 1; n <= NCoef; n++)
     {
-        y[0] += (float)ACoef[n] * x[n] - (float)BCoef[n] * y[n];
+        y[0] += ACoef[n] * x[n] - BCoef[n] * y[n];
     }
     return y[0];
 }
 
 float float_api_filters_iir_gyro_z(float float_new_sample)
 {
-    double ACoef[NCoef+1] = {
-        0.02675265688403948700,
-        0.10701062753615795000,
-        0.16051594130423694000,
-        0.10701062753615795000,
-        0.02675265688403948700
+    float ACoef[NCoef+1] = {
+        0.02675265688403948700f,
+        0.10701062753615795000f,
+        0.16051594130423694000f,
+        0.10701062753615795000f,
+        0.02675265688403948700f
     };
 
-    double BCoef[NCoef+1] = {
-        1.00000000000000000000,
-        -1.28555068801461640000,
-        1.01289917136114170000,
-        -0.35368582226312562000,
-        0.05438080336347629600
+    float BCoef[NCoef+1] = {
+        1.00000000000000000000f,
+        -1.28555068801461640000f,
+        1.01289917136114170000f,
+        -0.35368582226312562000f,
+        0.05438080336347629600f
     };
-
-    static float y[NCoef+1]; /* output samples*/
-    static float x[NCoef+1]; /* input samples */
+    
+    /* output samples*/
+    static float y[NCoef+1]; 
+    /* input samples */
+    static float x[NCoef+1]; 
     int n;
 
     /* shift the old samples */
-    for(n=NCoef; n>0; n--)
+    for(n = NCoef; n>0; n--)
     {
        x[n] = x[n-1];
        y[n] = y[n-1];
@@ -279,14 +294,13 @@ float float_api_filters_iir_gyro_z(float float_new_sample)
 
     /* Calculate the new output */
     x[0] = float_new_sample;
-    y[0] = (float)ACoef[0] * x[0];
-    for(n=1; n<=NCoef; n++)
+    y[0] = ACoef[0] * x[0];
+    for(n = 1; n <= NCoef; n++)
     {
-        y[0] += (float)ACoef[n] * x[n] - (float)BCoef[n] * y[n];
+        y[0] += ACoef[n] * x[n] - BCoef[n] * y[n];
     }
     return y[0];
 }
-
